requirements.txt
# Required libraries
twitchio
python-dotenv
requests
mysql-connector-python